+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+0.1.2 (2017-06-26)
+------------------
+
+* Add create multi-layers specimen.
+* Add number of layers X, Y, and Z as parameters.
+
 0.1.0 (2017-03-07)
 ------------------